backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.39
asyncmy==0.2.10
alembic==1.15.1
pydantic-settings==2.8.1
email-validator==2.2.0
python-dotenv==1.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.28.1
redis==5.2.1
celery==5.4.0
orjson==3.10.15


